@@ -1,28 +1,27 @@
 import { Layout, Menu } from 'antd';
-import basic from 'basic-reducers';
-import * as React from 'react';
-import { Route } from 'react-router';
-import { ReduxContainer } from "simple-redux-container";
-import { sevenReducer as seven } from "../../util/Seven.redux";
+import React from 'react';
+import { Route, Routes } from 'react-router';
+import { BrowserRouter } from 'react-router-dom';
 import { NewGame } from '../NewGame';
-import { Overview } from "../Overview";
-import { SetBids } from "../SetBids";
-import './App.css';
+import { Overview } from '../Overview';
+import { SetBids } from '../SetBids';
+import './App.scss';
+import 'antd/dist/antd.css';
 
 export const App = () => 
-  <ReduxContainer reducers={{basic, seven}}>
-    <Layout>
+    <Layout className="app">
       <Layout.Header>
         <Menu theme="dark" mode="horizontal">
-          <Menu.Item>Seven</Menu.Item>
+          <Menu.Item>Andy's Awesome Seven Scoring App</Menu.Item>
         </Menu>
       </Layout.Header>
       <Layout.Content>
-        <Route exact path="/" component={NewGame} />
-        <Route exact path="/overview" component={Overview} />
-        <Route exact path="/set-bids/:roundId" component={SetBids} />
+          <BrowserRouter>
+            <Routes>
+                <Route path="/" element={<NewGame />} />
+                <Route path="/overview" element={<Overview />} />
+                <Route path="/set-bids/:roundId" element={<SetBids />} />
+            </Routes>
+        </BrowserRouter>
       </Layout.Content>
-    </Layout>
-  </ReduxContainer>;
-
-export default App;
+    </Layout>;
