@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import App from './App.component';
+import {App} from './';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
