@@ -1,12 +1,16 @@
-import {Button, Icon, Slider, Tag} from 'antd';
-import * as React from 'react';
-import {HandCount, Player} from "../..//util/Seven.types";
-import {Col, Divider, Row} from "../Layout";
-import { SetBidsComponentProps } from './SetBids.types';
+import { ArrowLeftOutlined, ArrowRightOutlined, ArrowUpOutlined, LockOutlined, UnlockOutlined } from '@ant-design/icons';
+import { Button, Col, Divider, Row, Slider, Tag } from 'antd';
+import React from 'react';
+import { useNavigate } from 'react-router';
+import { HandCount, IPlayer } from '../../util/types';
+import { GotMineButton } from '../GotMineButton';
+import { GotScrewedButton } from '../GotScrewedButton';
+import { IBidOptionList, SetBidsProps } from "./SetBids.d";
+import './SetBids.scss';
 
-const getOptions = (handCount:HandCount, isDealer:boolean, cantBid:number):{} => {
-    const options = {[-1]: ''};
-    for(let i=0; i<=handCount; i++) {
+const getOptions = (handCount:HandCount, isDealer:boolean, cantBid:number):IBidOptionList => {
+    const options:IBidOptionList = {[-1]: ''};
+    for(let i=0; i<=handCount + 1; i++) {
         options[i] = !isDealer || i !== cantBid
             ? `${i}`
             : {
@@ -17,82 +21,85 @@ const getOptions = (handCount:HandCount, isDealer:boolean, cantBid:number):{} =>
     return options;
 };
 
-export const SetBidsComponent = (props:SetBidsComponentProps):JSX.Element => {
+export const SetBidsComponent = (props:SetBidsProps):JSX.Element => {
     const [bidsLocked, setBidsLocked] = React.useState(false);
     const lockBids = () => {setBidsLocked(true);}
     const unlockBids = () => {setBidsLocked(false);}
 
-    return <Row>
+    const round = props.round.current;
+
+    let players = [...props.players.list];
+    while(players[players.length - 1].id !== props.dealer.current(round.id)) {
+        players = players.slice(1).concat(players[0]);
+    }
+
+    const curDealer = props.dealer.current(round.id);
+
+    const navigate = useNavigate();
+    const nextRound = () => {navigate(`/set-bids/${round.id + 1}`);}
+    const prevRound = () => {navigate(`/set-bids/${round.id - 1}`);}
+    const backToOverview = () => {navigate('/overview');}
+
+    return <div className="set-bids">{!!props.round && <Row>
         <Col xs={24}>
-            <b>Round {props.round.id + 1} ({props.round.handCount} {props.round.handCount > 1 ? "hands" : "hand"})</b>
+            <b>Round {round.id + 1} ({round.handCount} {round.handCount > 1 ? "hands" : "hand"})</b>
         </Col>
         <Divider />
-        {props.players.map((player:Player, index:number) => 
+        {players.map((player:IPlayer, index:number) => 
             <Col xs={24} key={player.id}>
                 {player.name}&nbsp;
                 {index === 0 && <Tag color="green">Leads</Tag>}
-                {player.id === props.dealerId &&
+                {player.id === curDealer &&
                     <>
                         <Tag color="green">Dealer</Tag>
-                        {props.dealerCantBid >= 0 &&
-                            <Tag color="red">Can't bid {props.dealerCantBid}</Tag>
+                        {props.round.dealerCantBid >= 0 && !bidsLocked &&
+                            <Tag color="red">Can't bid {props.round.dealerCantBid}</Tag>
                         }
                     </>
                 }
-                <div style={{float: "right", display: props.canStart && bidsLocked ? "block" : "none"}}>
-                    <Tag
-                        color={props.getWonColor(player.id)}
-                        onClick={props.setHandStatus(props.round.id, player.id, props.getBid(player.id), true)}
-                    >
-                        Got my&nbsp;{props.getBid(player.id)}!
-                    </Tag>
-                    <Tag
-                        color={props.getLostColor(player.id)}
-                        onClick={props.setHandStatus(props.round.id, player.id, props.getBid(player.id), false)}
-                    >
-                        Got screwed
-                    </Tag>
+                <div style={{float: "right", display: bidsLocked ? "block" : "none"}}>
+                    <GotMineButton roundId={round.id} playerId={player.id} bid={props.bids.get.one(round.id, player.id)?.bid}/>
+                    <GotScrewedButton roundId={round.id} playerId={player.id} />
                 </div>
                 <div style={{marginBottom: bidsLocked ? "16px" : "32px"}}>
                     <Slider
                         disabled={bidsLocked}
                         min={-1}
-                        max={props.round.handCount}
-                        marks={getOptions(props.round.handCount, player.id === props.dealerId, props.dealerCantBid)}
+                        max={round.handCount}
+                        marks={getOptions(round.handCount, player.id === curDealer, props.round.dealerCantBid)}
                         step={1}
-                        value={props.getBid(player.id)}
-                        onChange={props.setBid(props.round.id, player.id)}
+                        value={props.bids.get.one(round.id, player.id)?.bid}
+                        onChange={props.bids.set(round.id, player.id) as (bid:number) => void}
                         style={{display: bidsLocked ? "none" : undefined}}
                     />
                 </div>
                 <Divider />
             </Col>
         )}
-        <Col xs={8}>
-            {props.round.id > 0 && <Button onClick={props.prevRound}>
-                <Icon type="arrow-left" /> Round&nbsp;{props.round.id}
+        <Col xs={6}>
+            {round.id > 0 && <Button onClick={prevRound} title='Previous round'>
+                <ArrowLeftOutlined />
             </Button>}
         </Col>
-        <Col xs={8}>
+        <Col xs={12} className="lock-btn-container">
             {!bidsLocked && <Button onClick={lockBids}>
-                <Icon type="unlock" /> Lock bids
+                <UnlockOutlined /> Lock bids
             </Button>}
             {bidsLocked && <Button onClick={unlockBids}>
-                <Icon type="lock" /> Unlock bids
+                <LockOutlined /> Unlock bids
             </Button>}
         </Col>
-        <Col xs={8}>
+        <Col xs={6}>
             <div style={{textAlign: "right"}}>
-                {props.round.id < 12 && <Button disabled={!props.canFinish} onClick={props.nextRound}>
-                    Round&nbsp;{props.round.id + 2} <Icon type="arrow-right" />
+                {round.id < 12 && <Button disabled={!props.round.canFinish} onClick={nextRound} title='Next round'>
+                    <ArrowRightOutlined />
                 </Button>}
             </div>
         </Col>
-        <Divider />
-        <Col xs={24}>
-            <Button disabled={!props.canFinish} onClick={props.backToOverview} style={{marginLeft: "25%", width: "50%"}}>
-                <Icon type="arrow-up" /> Back to overview
+        <Col xs={24} style={{textAlign: "center", marginTop: "16px"}}>
+            <Button onClick={backToOverview}>
+                <ArrowUpOutlined /> Back to overview
             </Button>
         </Col>
-    </Row>;
+    </Row>}</div>;
 }
